+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Maps `unknown_error` span status to `internal_error`. ([#5202](https://github.com/getsentry/relay/pull/5202))
 - Add `otp` and `two[-_]factor` to default scrubbing rules. ([#5250](https://github.com/getsentry/relay/pull/5250))
 - Add event merging logic for Playstation crashes. ([#5228](https://github.com/getsentry/relay/pull/5228))
+- Implement PII scrubbing for V2 spans. ([#5168](https://github.com/getsentry/relay/pull/5168))
 
 **Bug Fixes**: